Start view/model split
